requirements.txt
paho_mqtt==1.6.1
pyTelegramBotAPI==4.10.0
